refactor(capabilities): make text a submodule that owns its kinds
Collapse `text.rs` into a `text/` directory module split along its seams
and move the `aether.text` mail kinds out of `aether-kinds` so the
capability owns its own contract (ADR-0121). Pure code movement — no
behavior change, identical public surface, unchanged wire ids.

- `text.rs` -> `text/{mod,layout,kinds}` (beside the existing
  `text/atlas.rs`): `mod.rs` keeps the `#[bridge] mod native` actor +
  parked-request flow + handler dispatch; `layout.rs` holds the pure
  glyph-rasterization + immediate-mode-layout helpers (`text-native`
  gated); `kinds.rs` holds the mail kinds (always-on, wasm-safe).
- Move `LoadFont`, `LoadFontResult`, `DrawText`, `FontMetricsRequest`,
  `FontMetricsResult`, and `FontRef` from `aether-kinds` into
  `text/kinds.rs`, re-exported as `aether_capabilities::text::*`. Their
  five `aether.text.*` NAME-stability + codec round-trip tests move with
  them. `FontMetrics` / `GlyphAdvance` stay central in `aether-kinds`
  (consumed by its wasm-safe `text_metrics` primitive — moving them would
  form a crate cycle); the moved kinds reference them via
  `use aether_kinds::{FontMetrics, QuadSpace}`.
- Make the latent `ui -> text` vocabulary dependency explicit
  (`ui = ["text"]`); update `ui.rs` and the bundle test imports to pull
  the moved kinds from `aether_capabilities::text`.
