+++ crates/aether-capabilities/src/ui.rs
@@ -29,14 +29,14 @@ mod native {
 
     use aether_actor::actor;
     use aether_data::MailboxId;
-    use aether_kinds::{DrawSolidQuads, DrawText, QuadSpace, SolidQuad, UiClicked};
+    use aether_kinds::{DrawSolidQuads, QuadSpace, SolidQuad, UiClicked};
     use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx};
     use aether_substrate::chassis::error::BootError;
 
     use crate::input::{InputCapability, InputMailboxExt};
     use crate::lifecycle::{LifecycleCapability, LifecycleMailboxExt};
     use crate::render::RenderCapability;
-    use crate::text::TextCapability;
+    use crate::text::{DrawText, TextCapability};
 
     use super::{MouseButton, MouseMove, Tick, UiBar, UiButton, UiLabel, UiPanel};
 
@@ -283,8 +283,9 @@ mod native {
         use std::sync::mpsc::Receiver;
         use std::time::Duration;
 
+        use crate::text::DrawText;
         use aether_data::{Kind, MailId, MailboxId, SessionToken, Source, SourceAddr, Uuid};
-        use aether_kinds::{DrawSolidQuads, DrawText};
+        use aether_kinds::DrawSolidQuads;
         use aether_substrate::actor::native::binding::NativeBinding;
         use aether_substrate::mail::outbound::EgressEvent;
